+++ scripts/components/registry_client.ts
@@ -4,22 +4,22 @@ import { EOL } from 'os';
 import * as path from 'path';
 
 /**
- * Type for the response of `npm show <package> --json`
+ * Type for the response of `yarn npm info <package> --json`
  * We can add to this as we need to access additional config in a type-safe way
  */
 export type PackageInfo = {
-  // we have to match the output payload of npm show
+  // we have to match the output payload of yarn npm info
   // eslint-disable-next-line @typescript-eslint/naming-convention
   'dist-tags': Record<string, string>;
   deprecated?: string;
 };
 
 /**
- * Client for programmatically interacting with the local npm cli.
+ * Client for programmatically interacting with the local yarn cli.
  *
  * Note that this class is not guaranteed to be a singleton so it should not store any mutable internal state
  */
-export class NpmClient {
+export class RegistryClient {
   /**
    * execaCommand that allows us to capture stdout
    *
@@ -34,7 +34,7 @@ export class NpmClient {
   private readonly execWithIO;
 
   /**
-   * Initialize the npm client with an optional directory to operate in.
+   * Initialize the registry client with an optional directory to operate in.
    *
    * By default the client operates in the process cwd
    */
@@ -51,39 +51,39 @@ export class NpmClient {
     deprecationMessage: string,
   ) => {
     await this
-      .execWithIO`npm deprecate ${packageVersionSpecifier} ${deprecationMessage}`;
+      .execWithIO`yarn npm deprecate ${packageVersionSpecifier} ${deprecationMessage}`;
   };
 
   unDeprecatePackage = async (packageVersionSpecifier: string) => {
     // explicitly specifying an empty deprecation message is the official way to "un-deprecate" a package
     // see https://docs.npmjs.com/cli/v8/commands/npm-deprecate
-    await this.execWithIO`npm deprecate ${packageVersionSpecifier} ${''}`;
+    await this.execWithIO`yarn npm deprecate ${packageVersionSpecifier} ${''}`;
   };
 
   setDistTag = async (packageVersionSpecifier: string, distTag: string) => {
     await this
-      .execWithIO`npm dist-tag add ${packageVersionSpecifier} ${distTag}`;
+      .execWithIO`yarn npm tag add ${packageVersionSpecifier} ${distTag}`;
   };
 
   getPackageInfo = async (packageVersionSpecifier: string) => {
     const { stdout: jsonString } = await this
-      .exec`npm show ${packageVersionSpecifier} --json`;
+      .exec`yarn npm info ${packageVersionSpecifier} --json`;
     return JSON.parse(jsonString) as PackageInfo;
   };
 
   init = async () => {
-    await this.execWithIO`npm init --yes`;
+    await this.execWithIO`yarn init --yes`;
   };
 
   initWorkspacePackage = async (packageName: string) => {
-    await this.execWithIO`npm init --workspace packages/${packageName} --yes`;
+    await this.execWithIO`yarn init --workspace packages/${packageName} --yes`;
   };
 
   install = async (
     packageSpecifiers: string[],
     options: { dev: boolean } = { dev: false },
   ) => {
-    await this.execWithIO`npm install ${
+    await this.execWithIO`yarn add ${
       options.dev ? '-D' : ''
     } ${packageSpecifiers.join(' ')}`;
   };